Refactor : 예약내역 배치 기능 수정
