+++ src/main/java/com/project/team11_tabling/domain/booking/entity/Booking.java
@@ -15,16 +15,12 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.SQLDelete;
-import org.hibernate.annotations.SQLRestriction;
 
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Entity
-@SQLDelete(sql = "update booking set deleted_at = NOW() where id = ?")
-@SQLRestriction(value = "deleted_at is NULL")
 @Table(name = "booking")
 public class Booking extends Timestamp {
